+++ osversion.h
@@ -7,7 +7,7 @@ namespace mmh
 		bool is_windows_7_or_newer();
 		bool is_windows_vista_or_newer();
 		bool is_windows_xp_or_newer();
-		bool test_osversion(DWORD major, DWORD minor, DWORD sp = 0);
+		bool test_osversion(DWORD major, DWORD minor, WORD sp = 0);
 	};
 
 }
